+++ src/Enum/PrintSpeed.php
@@ -1,0 +1,19 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Janisvepris\ZplBuilder\Enum;
+
+enum PrintSpeed: string
+{
+    case Ips10 = '10';
+    case Ips11 = '11';
+    case Ips12 = '12';
+    case Ips2 = '2';
+    case Ips3 = '3';
+    case Ips4 = '4';
+    case Ips5 = '5';
+    case Ips6 = '6';
+    case Ips8 = '8';
+    case Ips9 = '9';
+}